Fix contextual panels breaking inside accordion
Styles for default panel were leaking into panels in other states (such as
success or warning). This fix constrains the default style and adds additional
selectors for the other states.

Since there is a visual overlap of borders in the according, a hierarchy
(danger > warning > success > primary) is enforced. For example the top and
bottom borders for a .panel-danger will take precedence over all other panel
types.

The test is added under panels.